# 09_order_food/docs/swagger.yml
swagger: '2.0'
info:
  title: Demo API
  description: Happy to code Demo API
  version: 1.0.0
schemes:
  - http
  - https
basePath: /
produces:
  - application/json
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
  /auth/register:
    post:
      summary: Register new user
      description: |
        Register new user
      parameters:
        - name: body
          in: body
          description: json body data to add new user
          required: true
          schema:
            required:
              - email
              - password
            type: object
            description: User Model
            properties:
              email:
                type: string
                description: email
              password:
                type: string
                description: password
              firstname:
                type: string
                description: firstname
              lastname:
                type: string
                description: lastname
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Auth
  /auth/login:
    post:
      summary: Login APi
      description: |
        Login API
      parameters:
        - name: body
          in: body
          description: json body data to login
          required: true
          schema:
            required:
              - email
              - password
            type: object
            description: Login Model
            properties:
              email:
                type: string
                description: email
              password:
                type: string
                description: password
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Auth

  /api/v1/users/list/{page}/{limit}:
    get:
      summary: Get all Users APi
      description: |
        Get all users API
      parameters:
        - name: authorization
          in: header
          description: json header data to login
          required: true
          type: string
        - in: path
          name: page
          type: integer
          description: Page
        - in: path
          name: limit
          type: integer
          description: Limit
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Users
  /api/v1/get/{id}:
    get:
      summary: Get all Users APi
      description: |
        Get all users API
      parameters:
        - name: authorization
          in: header
          description: json header data to login
          required: true
          type: string
        - in: path
          name: id
          type: integer
          description: User ID
      responses:
        '200':
          description: OK
        '400':
          description: Bad Request
        '401':
          description: Unauthorized
        '406':
          description: Not Acceptable
        '500':
          description: Server Internal error
      tags:
        - Users